// src/common/interfaces/place-detail-google.interface.ts
export interface PlaceDetailByGoogle {
  id: any;
  name: any;
  types: any;
  displayName: any;
  nationalPhoneNumber: any;
  formattedAddress: any;
  location: any;
  regularOpeningHours: { weekdayDescriptions: any };
  primaryTypeDisplayName: any;
  addressComponents: any;
  websiteUrl: any;
  parkingOptions: any;
  allowsDogs: any;
  goodForGroups: any;
  reservable: any;
  delivery: any;
  takeout: any;
}